CICO-65976: Reservation unresponsive after add to queue
The reservation screen becomes unresponsive after placing a room in queue
when the room key delivery setting is "No Key Delivery".

This issue was due to the failure of a check to skip the room key setup
for "No Key Delivery" because it was improperly retrieving the value of
the room key delivery setting.