CICO-75439: Remove vendor specific CRS setup partials
Removes the vendor specific Central Reservation System (CRS) setup
partials created for the addition of the rate type amount dropdown.
Instead of duplicating the setup page content, the rate type amount
dropdown has been added to the common setup partial and a check for
the existence of the rate_type_amount attribute in the settings
data determines whether the dropdown is rendered on the page.